+++ equitive_website/src/components/App.js
@@ -4,7 +4,7 @@ import Signup from './Signup';
 import { Container } from 'react-bootstrap'
 import { AuthProvider } from "../contexts/AuthContext"
 import { BrowserRouter as BrowserRouter, Routes, Route} from 'react-router-dom'
-import Dashboard from './Dashboard';
+import Dashboard from './Dashboard/Dashboard';
 import Login from './Login'
 import Landing from './LandingPage/Landing';
 import PrivateRoute from './PrivateRoute';
