+++ equitive_website/src/components/style.css
@@ -5,6 +5,10 @@ h2{
     font-family: 'Staatliches', cursive;
 }
 
+h1{
+    font-family: 'Staatliches', cursive;
+}
+
 .label{
     font-family: 'Inter', sans-serif;
 }
